Delete Todo
- Modify the route for authenticated only
- Prevent from deleting Todo by unauthorized ones

+++ app/Http/Controllers/TodosController.php
@@ -70,6 +70,10 @@ class TodosController extends Controller
      */
     public function destroy(Todo $todo)
     {
+        if ($todo->user_id !== auth()->user()->id) {
+            return response()->json('Unauthorized', 401);
+        }
+
         $todo->delete();
 
         return response('Deleted Todo item', 200);
